Refactor getFilenamePrefix into separate file
