+++ lighthouse-core/report/templates/report-template.html
@@ -246,37 +246,10 @@ limitations under the License.
 (function(exports) {
 'use strict';
 
-/**
- * Note: getFilenamePrefix is duplicated from asset-saver.js
- *
- * Generate a filenamePrefix of hostname_YYYY-MM-DD_HH-MM-SS
- * Date/time uses the local timezone, however Node has unreliable ICU
- * support, so we must construct a YYYY-MM-DD date format manually. :/
- * @param {!Object} results
- * @returns string
- */
-function getFilenamePrefix(results) {
-  const hostname = new URL(results.url).hostname;
-  const date = (results.generatedTime && new Date(results.generatedTime)) || new Date();
-
-  const timeStr = date.toLocaleTimeString('en-US', {hour12: false});
-  const dateParts = date.toLocaleDateString('en-US', {
-    year: 'numeric', month: '2-digit', day: '2-digit'
-  }).split('/');
-  dateParts.unshift(dateParts.pop());
-  const dateStr = dateParts.join('-');
-
-  const filenamePrefix = `${hostname}_${dateStr}_${timeStr}`;
-  // replace characters that are unfriendly to filenames
-  return filenamePrefix.replace(/[\/\?<>\\:\*\|":]/g, '-');
-}
-
 exports.addEventListener('DOMContentLoaded', _ => {
   exports.logger = new Logger('#log');
   new LighthouseReport(exports.lhresults);
 });
-
-exports.getFilenamePrefix = getFilenamePrefix;
 
 })(window);
 </script>
